requirements.txt
black==23.3.0
isort==5.12.0
Flask==2.2.3
pytest==7.2.2
psycopg2==2.9.5
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.8
python-dotenv==1.0.0
awscli==1.27.110
pandas==2.0.0
requests==2.28.2
beautifulsoup4==4.12.2
boto3==1.26.110
pydantic==1.10.7